Move websocket related code into ferrum-websocket crate
