+++ ferrum-shared/Cargo.toml
@@ -9,3 +9,8 @@ edition = "2018"
 serde = "1.0"
 chrono = { version = "0.4", features = ["serde"] }
 uuid = { version = "0.8", features = ["v4", "serde"] }
+actix-web = "4.0.0-beta.9"
+actix-http = "3.0.0-beta.10"
+jsonwebtoken = "7.2"
+futures = "0.3"
+thiserror = "1.0.24"
