+++ ferrum-shared/src/lib.rs
@@ -1,4 +1,18 @@
 pub mod channels;
+pub mod jwt;
 pub mod messages;
 pub mod servers;
 pub mod users;
+
+pub fn error_chain_fmt(
+    e: &impl std::error::Error,
+    f: &mut std::fmt::Formatter<'_>,
+) -> std::fmt::Result {
+    writeln!(f, "{}\n", e)?;
+    let mut current = e.source();
+    while let Some(cause) = current {
+        writeln!(f, "Caused by:\n\t{}", cause)?;
+        current = cause.source();
+    }
+    Ok(())
+}
