+++ ferrum-websocket/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "ferrum-websocket"
+version = "0.1.0"
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ferrum-shared = { path = "../ferrum-shared" }
+ferrum-db = { path = "../ferrum-db" }
+actix = "0.12"
+actix-web = "4.0.0-beta.9"
+actix-web-actors = "4.0.0-beta.7"
+sqlx = { version = "0.5.1", default-features = false, features = ["runtime-actix-rustls", "macros", "postgres", "uuid", "chrono", "migrate", "offline"] }
+uuid = { version = "0.8", features = ["v4", "serde"] }
+serde = "1.0"
+serde_json = "1.0.61"
