+++ src/routes/mod.rs
@@ -4,9 +4,7 @@ mod login;
 mod register;
 pub mod servers;
 pub mod users;
-pub mod websocket;
 
 pub use health_check::*;
 pub use login::*;
 pub use register::*;
-pub use websocket::websocket;
