+++ src/routes/servers/get_channels.rs
@@ -2,11 +2,10 @@ use actix_http::StatusCode;
 use actix_web::{web, HttpResponse, ResponseError};
 use anyhow::Context;
 use ferrum_db::{channels::queries::get_channels_for_server, users::queries::is_user_on_server};
-use ferrum_shared::channels::ChannelResponse;
+pub use ferrum_shared::error_chain_fmt;
+use ferrum_shared::{channels::ChannelResponse, jwt::AuthorizationService};
 use sqlx::PgPool;
 use uuid::Uuid;
-
-use crate::{error_chain_fmt, jwt::AuthorizationService};
 
 #[derive(thiserror::Error)]
 pub enum GetChannelsError {
