--- src/websocket/mod.rs
@@ -1,4 +1,0 @@
-pub mod messages;
-mod server;
-
-pub use server::*;
