+++ tests/api/servers/create_channel.rs
@@ -1,5 +1,5 @@
 use actix_http::{encoding::Decoder, Payload};
-use ferrum::websocket::messages::WebSocketMessage;
+use ferrum_websocket::messages::WebSocketMessage;
 use uuid::Uuid;
 
 use crate::helpers::{get_next_websocket_message, send_websocket_message, TestApplication};
